refactor: rename login to index default page
